Animate home drawer entries
